requirements.txt
langchain
streamlit
faiss-cpu 
langchain_google_genai
PyPDF2
langchain_openai 
langchain_core
python-dotenv
langchain_community
langserve
fastapi
uvicorn
sse_starlette
pypdf
groq
cassio
langchain-groq
langchainhub
beautifulsoup4
bs4
sentence-transformers
pdfplumber
langchain-pinecone
langchain-experimental
neo4j
yfiles_jupyter_graphs
wikipedia
selenium
unstructured




# unstructured
# tiktoken
# pinecone-client
# pypdf
# openai
# pandas 
# numpy
# python-dotenv
# google-generativeai
# langchain-pinecone

# wikipedia
# arxiv